+++ onsiteprint-plugin/blocks/event-creation/block-template-parts/block-form/steps/step-3.php
@@ -1,8 +1,8 @@
 <?php
 /* ------------------------------------------------------------------------
  *  Block Part Name: Step 3
- ?  Updated: 2025-01-02 - 02:52 (Y:m:d - H:i)
- ?  Info: Added new Title Text to the Dropdown Menu Button in Step 3 (Event Creation).
+ ?  Updated: 2025-01-09 - 04:55 (Y:m:d - H:i)
+ ?  Info: Added new Error Modal Template in Step 3 (Event Creation).
 ---------------------------------------------------------------------------
  #  The Block Part Content
 --------------------------------------------------------------------------- */
@@ -78,6 +78,19 @@
                 </div>
                 <div class="op-flex-row">
                     <iframe title="pdf" src="" style="min-height: 40rem; width: 100%"></iframe>
+                </div>
+            </template>
+            
+            <template id="<?= esc_attr($id) ?>-modal-error-template">
+                <div class="op-modal-header">
+                    <h3 class="op-modal-title"><?= esc_attr( $acf['step_3_error_title'] ) ?></h3>
+                    <button type="button" onclick="opToggleActive( 'class', 'op-modal ' ), opToggleActive( 'class', 'wp-block-post-content', 'op-modal-active' )" class="op-button-cancel op-button op-button-size-small op-button-style-outline" data-color="primary-90" data-icon="xmark" data-icon-position="right" data-title-visibility="1">
+                        <span class="op-icon" role="img" aria-label="X Mark Icon"></span>
+                        <span class="op-button-title"><?= esc_attr( $acf['step_3_cancel_button'] ) ?></span>
+                    </button>
+                </div>
+                <div class="op-flex-row">
+                    <p class="op-modal-description"><?= esc_attr( $acf['step_3_error_description'] ) ?></p>
                 </div>
             </template>